# perf(node/buffer): improve utf8 decoding performance