Microbenchmarks show 29x improvement of decoding uint8array

```
cpu: Apple M1 Max
runtime: deno 1.28.3 (aarch64-apple-darwin)

benchmark               time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------- -----------------------------
• bench
---------------------------------------------------------- -----------------------------
buf.toString         55.06 µs/iter  (52.58 µs … 240.75 µs)  53.46 µs 105.67 µs 109.08 µs
textdecoder length     1.9 µs/iter     (1.88 µs … 2.02 µs)   1.91 µs   2.02 µs   2.02 µs

summary for bench
  textdecoder length
   28.92x faster than buf.toString
```

<details>
<summary>Benchmark implementation</summary>

```javascript
const buf = Buffer.from(large)
const decoder = new TextDecoder()

group({ name: 'bench', summary: true }, () => {
  bench('buf.toString', () => {
    return buf.toString()
  })

  bench('textdecoder length', () => {
    return decoder.decode(large);
  })
})

await run({
  avg: true, // enable/disable avg column (default: true)
  json: false, // enable/disable json output (default: false)
  colors: true, // enable/disable colors (default: true)
  min_max: true, // enable/disable min/max column (default: true)
  // collect: false, // enable/disable collecting returned values into an array during the benchmark (default: false)
  // percentiles: false, // enable/disable percentiles column (default: true)
})
```

</details>