@anonrig 2022-12-06
I was wrong, this is 29x faster. Re-opening pull request